+++ packages/client/composables/useNav.ts
@@ -1,12 +1,18 @@
 import type { ClicksContext, SlideRoute, TocItem } from '@slidev/types'
-import type { ComputedRef, Ref, TransitionGroupProps } from 'vue'
+import type { ComputedRef, Ref, TransitionGroupProps, WritableComputedRef } from 'vue'
 import { computed, ref, watch } from 'vue'
-import type { Router } from 'vue-router'
+import { useRouter } from 'vue-router'
+import type { RouteLocationNormalized, Router } from 'vue-router'
+import { createSharedComposable } from '@vueuse/core'
+import { logicOr } from '@vueuse/math'
 import { getCurrentTransition } from '../logic/transition'
 import { getSlide, getSlidePath } from '../logic/slides'
 import { CLICKS_MAX } from '../constants'
 import { skipTransition } from '../logic/hmr'
+import { configs } from '../env'
+import { useRouteQuery } from '../logic/route'
 import { useTocTree } from './useTocTree'
+import { createClicksContextBase } from './useClicks'
 import { slides } from '#slidev/slides'
 
 export interface SlidevContextNav {
@@ -54,6 +60,27 @@ export interface SlidevContextNav {
   goLast: () => Promise<void>
 }
 
+export interface SlidevContextNavState {
+  router: Router
+  currentRoute: ComputedRef<RouteLocationNormalized>
+  isPrintMode: ComputedRef<boolean>
+  isPrintWithClicks: ComputedRef<boolean>
+  isEmbedded: ComputedRef<boolean>
+  isPlaying: ComputedRef<boolean>
+  isPresenter: ComputedRef<boolean>
+  isNotesViewer: ComputedRef<boolean>
+  isPresenterAvailable: ComputedRef<boolean>
+  hasPrimarySlide: ComputedRef<boolean>
+  currentSlideNo: ComputedRef<number>
+  currentSlideRoute: ComputedRef<SlideRoute>
+  clicksContext: ComputedRef<ClicksContext>
+  queryClicksRaw: Ref<string>
+  queryClicks: WritableComputedRef<number>
+  getPrimaryClicks: (route: SlideRoute) => ClicksContext
+}
+
+export interface SlidevContextNavFull extends SlidevContextNav, SlidevContextNavState {}
+
 export function useNavBase(
   currentSlideRoute: ComputedRef<SlideRoute>,
   clicksContext: ComputedRef<ClicksContext>,
@@ -96,7 +123,11 @@ export function useNavBase(
     return true
   }
 
-  const tocTree = useTocTree(slides)
+  const tocTree = useTocTree(
+    slides,
+    currentSlideNo,
+    currentSlideRoute,
+  )
 
   async function next() {
     clicksDirection.value = 1
@@ -201,3 +232,126 @@ export function useFixedNav(
     go: noop,
   }
 }
+
+const useNavState = createSharedComposable((): SlidevContextNavState => {
+  const router = useRouter()
+
+  const currentRoute = computed(() => router.currentRoute.value)
+  const isPrintMode = computed(() => currentRoute.value.query.print !== undefined)
+  const isPrintWithClicks = computed(() => currentRoute.value.query.print === 'clicks')
+  const isEmbedded = computed(() => currentRoute.value.query.embedded !== undefined)
+  const isPlaying = computed(() => currentRoute.value.name === 'play')
+  const isPresenter = computed(() => currentRoute.value.name === 'presenter')
+  const isNotesViewer = computed(() => currentRoute.value.name === 'notes')
+  const isPresenterAvailable = computed(() => !isPresenter.value && (!configs.remote || currentRoute.value.query.password === configs.remote))
+  const hasPrimarySlide = logicOr(isPlaying, isPresenter)
+
+  const currentSlideNo = computed(() => hasPrimarySlide.value ? getSlide(currentRoute.value.params.no as string)?.no ?? 1 : 1)
+  const currentSlideRoute = computed(() => slides.value[currentSlideNo.value - 1])
+
+  const queryClicksRaw = useRouteQuery<string>('clicks', '0')
+
+  const clicksContext = computed(() => getPrimaryClicks(currentSlideRoute.value))
+
+  const queryClicks = computed({
+    get() {
+      if (clicksContext.value.disabled)
+        return CLICKS_MAX
+      let v = +(queryClicksRaw.value || 0)
+      if (Number.isNaN(v))
+        v = 0
+      return v
+    },
+    set(v) {
+      queryClicksRaw.value = v.toString()
+    },
+  })
+
+  function getPrimaryClicks(
+    route: SlideRoute,
+  ): ClicksContext {
+    if (route?.meta?.__clicksContext)
+      return route.meta.__clicksContext
+
+    const thisNo = route.no
+    const context = createClicksContextBase(
+      computed({
+        get() {
+          if (context.disabled)
+            return CLICKS_MAX
+          if (currentSlideNo.value === thisNo)
+            return +(queryClicksRaw.value || 0) || 0
+          else if (currentSlideNo.value > thisNo)
+            return CLICKS_MAX
+          else
+            return 0
+        },
+        set(v) {
+          if (currentSlideNo.value === thisNo)
+            queryClicksRaw.value = Math.min(v, context.total).toString()
+        },
+      }),
+      route?.meta?.clicks,
+      () => isPrintMode.value && !isPrintWithClicks.value,
+    )
+
+    // On slide mounted, make sure the query is not greater than the total
+    context.onMounted = () => {
+      if (queryClicksRaw.value)
+        queryClicksRaw.value = Math.min(+queryClicksRaw.value, context.total).toString()
+    }
+
+    if (route?.meta)
+      route.meta.__clicksContext = context
+
+    return context
+  }
+
+  return {
+    router,
+    currentRoute,
+    isPrintMode,
+    isPrintWithClicks,
+    isEmbedded,
+    isPlaying,
+    isPresenter,
+    isNotesViewer,
+    isPresenterAvailable,
+    hasPrimarySlide,
+    currentSlideNo,
+    currentSlideRoute,
+    clicksContext,
+    queryClicksRaw,
+    queryClicks,
+    getPrimaryClicks,
+  }
+})
+
+export const useNav = createSharedComposable((): SlidevContextNavFull => {
+  const state = useNavState()
+  const router = useRouter()
+
+  const nav = useNavBase(
+    state.currentSlideRoute,
+    state.clicksContext,
+    state.queryClicks,
+    state.isPresenter,
+    router,
+  )
+
+  watch(
+    [nav.total, state.currentRoute],
+    async () => {
+      if (state.hasPrimarySlide.value && !getSlide(state.currentRoute.value.params.no as string)) {
+        // The current slide may has been removed. Redirect to the last slide.
+        await nav.goLast()
+      }
+    },
+    { flush: 'pre', immediate: true },
+  )
+
+  return {
+    ...nav,
+    ...state,
+  }
+})
