+++ packages/client/composables/useSwipeControls.ts
@@ -1,7 +1,7 @@
 import type { Ref } from 'vue'
 import { ref } from 'vue'
 import { timestamp, usePointerSwipe } from '@vueuse/core'
-import { useNav } from '../logic/nav'
+import { useNav } from '../composables/useNav'
 import { useDrawings } from './useDrawings'
 
 export function useSwipeControls(root: Ref<HTMLElement | undefined>) {
