+++ packages/client/internals/SlideContainer.vue
@@ -3,8 +3,7 @@ import { provideLocal, useElementSize, useStyleTag } from '@vueuse/core'
 import { computed, ref, watchEffect } from 'vue'
 import { configs, slideAspect, slideHeight, slideWidth } from '../env'
 import { injectionSlideScale } from '../constants'
-import { useNav } from '../logic/nav'
-import { useNavState } from '../logic/nav-state'
+import { useNav } from '../composables/useNav'
 
 const props = defineProps({
   width: {
@@ -22,8 +21,7 @@ const props = defineProps({
   },
 })
 
-const { clicksDirection } = useNav()
-const { isPrintMode } = useNavState()
+const { clicksDirection, isPrintMode } = useNav()
 
 const root = ref<HTMLDivElement>()
 const element = useElementSize(root)
