+++ packages/client/setup/shortcuts.ts
@@ -3,7 +3,7 @@ import type { NavOperations, ShortcutOptions } from '@slidev/types'
 import { downloadPDF } from '../utils'
 import { toggleDark } from '../logic/dark'
 import { magicKeys, showGotoDialog, showOverview, toggleOverview } from '../state'
-import { useNav } from '../logic/nav'
+import { useNav } from '../composables/useNav'
 import { useDrawings } from '../composables/useDrawings'
 import { currentOverviewPage, downOverviewPage, nextOverviewPage, prevOverviewPage, upOverviewPage } from './../logic/overview'
 import setups from '#slidev/setups/shortcuts'
